+++ include/AstMath.h
@@ -1,0 +1,35 @@
+/*
+ * ASTU - AST Utilities
+ * A collection of Utilities for Applied Software Techniques (AST).
+ * 
+ * Copyright (c) 2020, 2021 Roman Divotkey, Nora Loimayr. All rights reserved.
+ * 
+ *----------------------------------------------------------------------------
+ * Meanings of ASTU: Slightly melted butter (Hinduism / Religious Law), 
+ * Let it be / Be it so (Sanskrit), City (Ancient Greek)
+ *----------------------------------------------------------------------------
+ */
+
+/**
+ * @file
+ * @brief This file defines public functions, templates and classes, offered by the mathematics module of AST-Utilities.
+ */
+
+#pragma once
+
+// ASTU Library includes
+#include "Vector2.h"
+#include "Matrix3.h"
+
+namespace astu {
+
+    /**
+     * @defgroup math_group Mathematics
+     * @brief This module contains classes and functions dedicated to graphics.
+     * @{
+     */
+
+    /**@}*/
+
+
+} // end of namespace